Revert "Revert "feat: Add option to use rocksdb record store (#2074)" (#2082)"
This reverts commit 66b7e672ba3a784757941986c80d51652079cc7f.
