+++ dozer-cli/src/simple/orchestrator.rs
@@ -5,7 +5,10 @@ use crate::pipeline::PipelineBuilder;
 use crate::shutdown::ShutdownReceiver;
 use crate::simple::build;
 use crate::simple::helper::validate_config;
-use crate::utils::{get_cache_manager_options, get_default_max_num_records, get_executor_options};
+use crate::utils::{
+    get_cache_manager_options, get_checkpoint_options, get_default_max_num_records,
+    get_executor_options,
+};
 
 use crate::{flatten_join_handle, join_handle_map_err};
 use dozer_api::auth::{Access, Authorizer};
@@ -226,7 +229,7 @@ impl SimpleOrchestrator {
             &self.config.sources,
             self.config.sql.as_deref(),
             &self.config.endpoints,
-            self.config.app.data_storage.clone(),
+            get_checkpoint_options(&self.config),
             self.labels.clone(),
             &self.config.udfs,
         ))?;
