+++ dozer-recordstore/src/in_memory/store.rs
@@ -3,33 +3,11 @@ use std::{
     sync::{Arc, Weak},
 };
 
-use dozer_types::{
-    bincode,
-    errors::internal::BoxedError,
-    parking_lot::RwLock,
-    serde::{Deserialize, Serialize},
-    thiserror::Error,
-    types::{Field, Lifetime, Record},
-};
+use dozer_types::{bincode, parking_lot::RwLock, types::Field};
 
-use crate::RecordRefInner;
+use crate::RecordStoreError;
 
-use super::{FieldRef, ProcessorRecord, RecordRef};
-
-#[derive(Error, Debug)]
-pub enum RecordStoreError {
-    #[error("Unable to deserialize type: {} - Reason: {}", typ, reason.to_string())]
-    DeserializationError {
-        typ: &'static str,
-        reason: BoxedError,
-    },
-
-    #[error("Unable to serialize type: {} - Reason: {}", typ, reason.to_string())]
-    SerializationError {
-        typ: &'static str,
-        reason: BoxedError,
-    },
-}
+use super::{RecordRef, RecordRefInner};
 
 pub trait StoreRecord {
     fn store_record(&self, record: &RecordRef) -> Result<(), RecordStoreError>;
@@ -37,32 +15,6 @@ pub trait StoreRecord {
     fn create_ref(&self, values: &[Field]) -> Result<RecordRef, RecordStoreError> {
         let record = RecordRef::new(values.to_vec());
         self.store_record(&record)?;
-        Ok(record)
-    }
-
-    fn load_ref<'a>(
-        &self,
-        record_ref: &'a RecordRef,
-    ) -> Result<Vec<FieldRef<'a>>, RecordStoreError> {
-        Ok(record_ref.load())
-    }
-
-    fn create_record(&self, record: &Record) -> Result<ProcessorRecord, RecordStoreError> {
-        let record_ref = self.create_ref(&record.values)?;
-        let mut processor_record = ProcessorRecord::new(Box::new([record_ref]));
-        processor_record.set_lifetime(record.lifetime.clone());
-        Ok(processor_record)
-    }
-
-    fn load_record(&self, processor_record: &ProcessorRecord) -> Result<Record, RecordStoreError> {
-        let mut record = Record::default();
-        for record_ref in processor_record.values.iter() {
-            let fields = self.load_ref(record_ref)?;
-            record
-                .values
-                .extend(fields.iter().map(|field| field.cloned()));
-        }
-        record.set_lifetime(processor_record.get_lifetime());
         Ok(record)
     }
 }
@@ -110,19 +62,6 @@ impl ProcessorRecordStore {
             .record_pointer_to_index
             .get(&(record_ref.id()))
             .expect("RecordRef not found in ProcessorRecordStore") as u64
-    }
-
-    pub fn serialize_record(&self, record: &ProcessorRecord) -> Result<Vec<u8>, bincode::Error> {
-        let ProcessorRecord { values, lifetime } = record;
-        let values = values
-            .iter()
-            .map(|value| self.serialize_ref(value))
-            .collect();
-        let record = ProcessorRecordForSerialization {
-            values,
-            lifetime: lifetime.clone(),
-        };
-        bincode::serialize(&record)
     }
 }
 
@@ -184,22 +123,16 @@ impl ProcessorRecordStoreDeserializer {
         Ok(())
     }
 
-    pub fn deserialize_ref(&self, index: u64) -> RecordRef {
-        self.inner.read().records[index as usize]
+    pub fn deserialize_ref(&self, index: u64) -> Result<RecordRef, RecordStoreError> {
+        Ok(self
+            .inner
+            .read()
+            .records
+            .get(index as usize)
+            .ok_or(RecordStoreError::InMemoryRecordNotFound(index))?
             .as_ref()
-            .unwrap_or_else(|| {
-                panic!("RecordRef {index} not found in ProcessorRecordStoreDeserializer")
-            })
-            .clone()
-    }
-
-    pub fn deserialize_record(&self, data: &[u8]) -> Result<ProcessorRecord, bincode::Error> {
-        let ProcessorRecordForSerialization { values, lifetime } = bincode::deserialize(data)?;
-        let values = values
-            .iter()
-            .map(|index| self.deserialize_ref(*index))
-            .collect();
-        Ok(ProcessorRecord { values, lifetime })
+            .ok_or(RecordStoreError::InMemoryRecordNotFound(index))?
+            .clone())
     }
 
     pub fn into_record_store(self) -> ProcessorRecordStore {
@@ -243,42 +176,9 @@ fn insert_record_pointer_to_index(
     debug_assert!(previous_index.is_none());
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(crate = "dozer_types::serde")]
-struct ProcessorRecordForSerialization {
-    values: Vec<u64>,
-    lifetime: Option<Box<Lifetime>>,
-}
-
 #[cfg(test)]
 mod tests {
-    use std::time::Duration;
-
-    use dozer_types::types::Timestamp;
-
     use super::*;
-
-    fn test_record() -> Record {
-        let mut record = Record::new(vec![
-            Field::Int(1),
-            Field::Int(2),
-            Field::Int(3),
-            Field::Int(4),
-        ]);
-        record.lifetime = Some(Lifetime {
-            reference: Timestamp::parse_from_rfc3339("2020-01-01T00:13:00Z").unwrap(),
-            duration: Duration::from_secs(10),
-        });
-        record
-    }
-
-    #[test]
-    fn test_record_roundtrip() {
-        let record = test_record();
-        let record_store = ProcessorRecordStore::new().unwrap();
-        let processor_record = record_store.create_record(&record).unwrap();
-        assert_eq!(record_store.load_record(&processor_record).unwrap(), record);
-    }
 
     #[test]
     fn test_serialization_roundtrip() {
@@ -304,21 +204,9 @@ mod tests {
         record_store.deserialize_and_extend(&data).unwrap();
         let mut deserialized_record_refs = vec![];
         for serialized_record_ref in serialized_record_refs {
-            deserialized_record_refs.push(record_store.deserialize_ref(serialized_record_ref));
+            deserialized_record_refs
+                .push(record_store.deserialize_ref(serialized_record_ref).unwrap());
         }
         assert_eq!(deserialized_record_refs, record_refs);
-    }
-
-    #[test]
-    fn test_record_serialization_roundtrip() {
-        let record_store = ProcessorRecordStore::new().unwrap();
-        let record = record_store.create_record(&test_record()).unwrap();
-        let serialized_record = record_store.serialize_record(&record).unwrap();
-        let data = record_store.serialize_slice(0).unwrap().0;
-
-        let record_store = ProcessorRecordStoreDeserializer::new().unwrap();
-        record_store.deserialize_and_extend(&data).unwrap();
-        let deserialized_record = record_store.deserialize_record(&serialized_record).unwrap();
-        assert_eq!(deserialized_record, record);
     }
 }
