+++ dozer-recordstore/src/lib.rs
@@ -1,449 +1,42 @@
-//! [`RecordRef`] is a compact representation of a collection of [dozer_types::types::Field]s
-//! There are two principles that make this representation more compact than `[Field]`:
-//!  1. The fields and their types are stored as a Struct of Arrays instead of
-//!     and Array of Structs. This makes it possible to pack the discriminants
-//!     for the field types as a byte per field, instead of taking up a full word,
-//!     which is the case in [Field] (because the variant value must be aligned)
-//!  2. The field values are stored packed. In a `[Field]` representation, each
-//!     field takes as much space as the largest enum variant in [Field] (plus its discriminant,
-//!     see (1.)). Instead, for the compact representation, we pack the values into
-//!     align_of::<Field>() sized slots. This way, a u64 takes only 8 bytes, whereas
-//!     a u128 can still use its 16 bytes.
-use std::alloc::{dealloc, handle_alloc_error, Layout};
-use std::sync::Arc;
-use std::{hash::Hash, ptr::NonNull};
-
-use slice_dst::SliceWithHeader;
-
-use dozer_types::chrono::{DateTime, FixedOffset, NaiveDate};
-use dozer_types::json_types::JsonValue;
-use dozer_types::ordered_float::OrderedFloat;
-use dozer_types::rust_decimal::Decimal;
-use dozer_types::types::{DozerDuration, DozerPoint};
 use dozer_types::{
+    bincode,
+    errors::internal::BoxedError,
+    models::app_config::RecordStore,
     serde::{Deserialize, Serialize},
-    types::{Field, FieldType, Lifetime},
+    thiserror::{self, Error},
+    types::{Field, Lifetime, Record},
 };
+use in_memory::{FieldRef, StoreRecord as _};
 
-// The alignment of an enum is necessarily the maximum alignment of its variants
-// (otherwise it would be unsound to read from it).
-// So, by using the alignment of `Field` as the alignment of the values in our
-// packed `RecordRef`, we ensure that all accesses are aligned.
-// This wastes a little bit of memory for subsequent fields that have
-// smaller minimum alignment and size (such as `bool`, which has size=1, align=1),
-// but in practice this should be negligible compared to the added effort of
-// packing these fields while keeping everything aligned.
-const MAX_ALIGN: usize = std::mem::align_of::<Field>();
+#[derive(Error, Debug)]
+pub enum RecordStoreError {
+    #[error("Unable to deserialize type: {} - Reason: {}", typ, reason.to_string())]
+    DeserializationError {
+        typ: &'static str,
+        reason: BoxedError,
+    },
 
-#[repr(transparent)]
-#[derive(Debug)]
-/// `repr(transparent)` inner struct so we can implement drop logic on it
-/// This is a `slice_dst` `SliceWithHeader` so we can make a fat Arc, saving a level
-/// of indirection and a pointer which would otherwise be needed for the field types
-struct RecordRefInner(SliceWithHeader<NonNull<u8>, Option<FieldType>>);
-
-unsafe impl Send for RecordRefInner {}
-unsafe impl Sync for RecordRefInner {}
-
-#[derive(Debug, Clone)]
-pub struct RecordRef(Arc<RecordRefInner>);
-
-impl PartialEq for RecordRef {
-    fn eq(&self, other: &Self) -> bool {
-        self.load() == other.load()
-    }
+    #[error("Unable to serialize type: {} - Reason: {}", typ, reason.to_string())]
+    SerializationError {
+        typ: &'static str,
+        reason: BoxedError,
+    },
+    #[error("Failed to create tempdir: {0}")]
+    FailedToCreateTempDir(#[source] std::io::Error),
+    #[error("Storage error: {0}")]
+    Storage(#[from] dozer_storage::errors::StorageError),
+    #[error("In memory record not found {0}")]
+    InMemoryRecordNotFound(u64),
+    #[error("Rocksdb record not found: {0}")]
+    RocksdbRecordNotFound(u64),
+    #[error("Bincode error: {0}")]
+    Bincode(#[from] bincode::Error),
 }
 
-impl Eq for RecordRef {}
-
-impl Hash for RecordRef {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.load().hash(state)
-    }
-}
-
-impl<'de> Deserialize<'de> for RecordRef {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: dozer_types::serde::Deserializer<'de>,
-    {
-        let fields = Vec::<FieldRef>::deserialize(deserializer)?;
-        let owned_fields: Vec<_> = fields.iter().map(FieldRef::cloned).collect();
-        Ok(Self::new(owned_fields))
-    }
-}
-impl Serialize for RecordRef {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: dozer_types::serde::Serializer,
-    {
-        self.load().serialize(serializer)
-    }
-}
-
-#[inline(always)]
-unsafe fn adjust_alignment<T>(ptr: *mut u8) -> *mut u8 {
-    ptr.add(ptr.align_offset(std::mem::align_of::<T>()))
-}
-/// # Safety
-/// ptr should be valid for writing a `T`,
-/// that is, ptr..ptr + size_of::<T> should be inside a single live allocation
-unsafe fn write<T>(ptr: *mut u8, value: T) -> *mut u8 {
-    let ptr = adjust_alignment::<T>(ptr) as *mut T;
-    ptr.write(value);
-    ptr.add(1) as *mut u8
-}
-
-/// # Safety
-/// ptr should be valid for reading a `T`,
-/// that is, ptr..ptr + size_of::<T> should be inside a single live allocation
-/// and the memory read should be initialized.
-/// The returned reference is only valid as long as pointed to memory is valid
-/// for reading.
-unsafe fn read_ref<'a, T>(ptr: *mut u8) -> (*mut u8, &'a T) {
-    let ptr = adjust_alignment::<T>(ptr) as *mut T;
-    let result = &*ptr;
-    (ptr.add(1) as *mut u8, result)
-}
-
-/// # Safety
-/// ptr should be valid for reading a `T`,
-/// that is, ptr..ptr + size_of::<T> should be inside a single live allocation
-/// and the memory read should be initialized.
-/// This takes ownership of the memory returned as `T`, which means dropping `T`
-/// may make future reads from `ptr` undefined behavior
-unsafe fn read<T>(ptr: *mut u8) -> (*mut u8, T) {
-    let ptr = adjust_alignment::<T>(ptr) as *mut T;
-    let result = ptr.read();
-    (ptr.add(1) as *mut u8, result)
-}
-
-/// # Safety
-/// `ptr` should be valid for reading the contents of a `Field` with the type
-/// corresponding to `field_type`.
-/// See `read_ref`
-unsafe fn read_field_ref<'a>(ptr: *mut u8, field_type: FieldType) -> (*mut u8, FieldRef<'a>) {
-    match field_type {
-        FieldType::UInt => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::UInt(*value))
-        }
-        FieldType::U128 => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::U128(*value))
-        }
-
-        FieldType::Int => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Int(*value))
-        }
-
-        FieldType::I128 => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::I128(*value))
-        }
-
-        FieldType::Float => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Float(*value))
-        }
-
-        FieldType::Boolean => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Boolean(*value))
-        }
-
-        FieldType::String => {
-            let (ptr, value): (_, &String) = read_ref(ptr);
-            (ptr, FieldRef::String(value))
-        }
-        FieldType::Text => {
-            let (ptr, value): (_, &String) = read_ref(ptr);
-            (ptr, FieldRef::Text(value))
-        }
-        FieldType::Binary => {
-            let (ptr, value): (_, &Vec<u8>) = read_ref(ptr);
-            (ptr, FieldRef::Binary(value))
-        }
-        FieldType::Decimal => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Decimal(*value))
-        }
-        FieldType::Timestamp => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Timestamp(*value))
-        }
-        FieldType::Date => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Date(*value))
-        }
-        FieldType::Json => {
-            let (ptr, value) = read_ref::<JsonValue>(ptr);
-            (ptr, FieldRef::Json(value.to_owned()))
-        }
-        FieldType::Point => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Point(*value))
-        }
-        FieldType::Duration => {
-            let (ptr, value) = read_ref(ptr);
-            (ptr, FieldRef::Duration(*value))
-        }
-    }
-}
-unsafe fn read_field(ptr: *mut u8, field_type: FieldType) -> (*mut u8, Field) {
-    match field_type {
-        FieldType::UInt => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::UInt(value))
-        }
-        FieldType::U128 => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::U128(value))
-        }
-
-        FieldType::Int => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Int(value))
-        }
-
-        FieldType::I128 => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::I128(value))
-        }
-
-        FieldType::Float => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Float(value))
-        }
-
-        FieldType::Boolean => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Boolean(value))
-        }
-
-        FieldType::String => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::String(value))
-        }
-        FieldType::Text => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::String(value))
-        }
-        FieldType::Binary => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Binary(value))
-        }
-        FieldType::Decimal => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Decimal(value))
-        }
-        FieldType::Timestamp => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Timestamp(value))
-        }
-        FieldType::Date => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Date(value))
-        }
-        FieldType::Json => {
-            let (ptr, value) = read::<JsonValue>(ptr);
-            (ptr, Field::Json(value))
-        }
-        FieldType::Point => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Point(value))
-        }
-        FieldType::Duration => {
-            let (ptr, value) = read(ptr);
-            (ptr, Field::Duration(value))
-        }
-    }
-}
-
-#[inline(always)]
-fn add_field_size<T>(size: &mut usize) {
-    let align = std::mem::align_of::<T>();
-    // Align the start of the field
-    *size = (*size + (align - 1)) & !(align - 1);
-    *size += std::mem::size_of::<T>();
-}
-fn size(fields: &[Option<FieldType>]) -> usize {
-    let mut size = 0;
-    for field in fields.iter().flatten() {
-        match field {
-            FieldType::UInt => add_field_size::<u64>(&mut size),
-            FieldType::U128 => add_field_size::<u128>(&mut size),
-            FieldType::Int => add_field_size::<i64>(&mut size),
-            FieldType::I128 => add_field_size::<i128>(&mut size),
-            FieldType::Float => add_field_size::<OrderedFloat<f64>>(&mut size),
-            FieldType::Boolean => add_field_size::<bool>(&mut size),
-            FieldType::String => add_field_size::<String>(&mut size),
-            FieldType::Text => add_field_size::<String>(&mut size),
-            FieldType::Binary => add_field_size::<Vec<u8>>(&mut size),
-            FieldType::Decimal => add_field_size::<Decimal>(&mut size),
-            FieldType::Timestamp => add_field_size::<DateTime<FixedOffset>>(&mut size),
-            FieldType::Date => add_field_size::<NaiveDate>(&mut size),
-            FieldType::Json => add_field_size::<JsonValue>(&mut size),
-            FieldType::Point => add_field_size::<DozerPoint>(&mut size),
-            FieldType::Duration => add_field_size::<DozerDuration>(&mut size),
-        }
-    }
-    size
-}
-
-#[derive(Hash, Serialize, Deserialize, Debug, PartialEq, Eq)]
-#[serde(crate = "dozer_types::serde")]
-pub enum FieldRef<'a> {
-    UInt(u64),
-    U128(u128),
-    Int(i64),
-    I128(i128),
-    Float(OrderedFloat<f64>),
-    Boolean(bool),
-    String(&'a str),
-    Text(&'a str),
-    Binary(&'a [u8]),
-    Decimal(Decimal),
-    Timestamp(DateTime<FixedOffset>),
-    Date(NaiveDate),
-    Json(JsonValue),
-    Point(DozerPoint),
-    Duration(DozerDuration),
-    Null,
-}
-
-impl FieldRef<'_> {
-    pub fn cloned(&self) -> Field {
-        match self {
-            FieldRef::UInt(v) => Field::UInt(*v),
-            FieldRef::U128(v) => Field::U128(*v),
-            FieldRef::Int(v) => Field::Int(*v),
-            FieldRef::I128(v) => Field::I128(*v),
-            FieldRef::Float(v) => Field::Float(*v),
-            FieldRef::Boolean(v) => Field::Boolean(*v),
-            FieldRef::String(v) => Field::String((*v).to_owned()),
-            FieldRef::Text(v) => Field::Text((*v).to_owned()),
-            FieldRef::Binary(v) => Field::Binary((*v).to_vec()),
-            FieldRef::Decimal(v) => Field::Decimal(*v),
-            FieldRef::Timestamp(v) => Field::Timestamp(*v),
-            FieldRef::Date(v) => Field::Date(*v),
-            FieldRef::Json(v) => Field::Json(v.clone()),
-            FieldRef::Point(v) => Field::Point(*v),
-            FieldRef::Duration(v) => Field::Duration(*v),
-            FieldRef::Null => Field::Null,
-        }
-    }
-}
-
-impl RecordRef {
-    pub fn new(fields: Vec<Field>) -> Self {
-        let field_types = fields
-            .iter()
-            .map(|field| field.ty())
-            .collect::<Box<[Option<FieldType>]>>();
-        let size = size(&field_types);
-
-        let layout = Layout::from_size_align(size, MAX_ALIGN).unwrap();
-        // SAFETY: Everything is `ALIGN` byte aligned
-        let data = unsafe {
-            let data = std::alloc::alloc(layout);
-            if data.is_null() {
-                handle_alloc_error(layout);
-            }
-            data
-        };
-        // SAFETY: We checked for null above
-        let data = unsafe { NonNull::new_unchecked(data) };
-        let mut ptr = data.as_ptr();
-
-        // SAFETY:
-        // - ptr is non-null (we got it from a NonNull)
-        // - ptr is dereferencable (its memory range is large enough and not de-allocated)
-        //
-        unsafe {
-            for field in fields {
-                match field {
-                    Field::UInt(v) => ptr = write(ptr, v),
-                    Field::U128(v) => ptr = write(ptr, v),
-                    Field::Int(v) => ptr = write(ptr, v),
-                    Field::I128(v) => ptr = write(ptr, v),
-                    Field::Float(v) => ptr = write(ptr, v),
-                    Field::Boolean(v) => ptr = write(ptr, v),
-                    Field::String(v) => ptr = write(ptr, v),
-                    Field::Text(v) => ptr = write(ptr, v),
-                    Field::Binary(v) => ptr = write(ptr, v),
-                    Field::Decimal(v) => ptr = write(ptr, v),
-                    Field::Timestamp(v) => ptr = write(ptr, v),
-                    Field::Date(v) => ptr = write(ptr, v),
-                    Field::Json(v) => ptr = write(ptr, v),
-                    Field::Point(v) => ptr = write(ptr, v),
-                    Field::Duration(v) => ptr = write(ptr, v),
-                    Field::Null => (),
-                }
-            }
-        }
-        // SAFETY: This is valid, because inner is `repr(transparent)`
-        let arc = unsafe {
-            let arc = SliceWithHeader::from_slice::<Arc<_>>(data, &field_types);
-            std::mem::transmute(arc)
-        };
-        Self(arc)
-    }
-
-    pub fn load(&self) -> Vec<FieldRef<'_>> {
-        self.0
-            .field_types()
-            .iter()
-            .scan(self.0.data().as_ptr(), |ptr, field_type| {
-                let Some(field_type) = field_type else {
-                    return Some(FieldRef::Null);
-                };
-
-                unsafe {
-                    let (new_ptr, value) = read_field_ref(*ptr, *field_type);
-                    *ptr = new_ptr;
-                    Some(value)
-                }
-            })
-            .collect()
-    }
-
-    #[inline(always)]
-    pub fn id(&self) -> usize {
-        Arc::as_ptr(&self.0) as *const () as usize
-    }
-}
-
-impl RecordRefInner {
-    #[inline(always)]
-    fn field_types(&self) -> &[Option<FieldType>] {
-        &self.0.slice
-    }
-
-    #[inline(always)]
-    fn data(&self) -> NonNull<u8> {
-        self.0.header
-    }
-}
-
-impl Drop for RecordRefInner {
-    fn drop(&mut self) {
-        let mut ptr = self.data().as_ptr();
-        for field in self.field_types().iter().flatten() {
-            unsafe {
-                // Read owned so all field destructors run
-                ptr = read_field(ptr, *field).0;
-            }
-        }
-        // Then deallocate the field storage
-        unsafe {
-            dealloc(
-                self.data().as_ptr(),
-                Layout::from_size_align(size(self.field_types()), MAX_ALIGN).unwrap(),
-            );
-        }
-    }
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum RecordRef {
+    InMemory(in_memory::RecordRef),
+    Rocksdb(u64),
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
@@ -483,53 +76,232 @@ impl ProcessorRecord {
     }
 }
 
-mod store;
-pub use store::{
-    ProcessorRecordStore, ProcessorRecordStoreDeserializer, RecordStoreError, StoreRecord,
-};
+pub trait StoreRecord {
+    fn create_ref(&self, values: &[Field]) -> Result<RecordRef, RecordStoreError>;
+
+    fn load_ref(&self, record_ref: &RecordRef) -> Result<Vec<Field>, RecordStoreError>;
+
+    fn create_record(&self, record: &Record) -> Result<ProcessorRecord, RecordStoreError> {
+        let record_ref = self.create_ref(&record.values)?;
+        let mut processor_record = ProcessorRecord::new(Box::new([record_ref]));
+        processor_record.set_lifetime(record.lifetime.clone());
+        Ok(processor_record)
+    }
+
+    fn load_record(&self, processor_record: &ProcessorRecord) -> Result<Record, RecordStoreError> {
+        let mut record = Record::default();
+        for record_ref in processor_record.values.iter() {
+            let fields = self.load_ref(record_ref)?;
+            record.values.extend(fields);
+        }
+        record.set_lifetime(processor_record.get_lifetime());
+        Ok(record)
+    }
+}
+
+#[derive(Debug)]
+pub enum ProcessorRecordStore {
+    InMemory(in_memory::ProcessorRecordStore),
+    Rocksdb(rocksdb::ProcessorRecordStore),
+}
+
+impl ProcessorRecordStore {
+    pub fn new(record_store: RecordStore) -> Result<Self, RecordStoreError> {
+        match record_store {
+            RecordStore::InMemory => Ok(Self::InMemory(in_memory::ProcessorRecordStore::new()?)),
+            RecordStore::Rocksdb => Ok(Self::Rocksdb(rocksdb::ProcessorRecordStore::new()?)),
+        }
+    }
+
+    pub fn num_records(&self) -> usize {
+        match self {
+            Self::InMemory(store) => store.num_records(),
+            Self::Rocksdb(store) => store.num_records(),
+        }
+    }
+
+    pub fn serialize_slice(&self, start: usize) -> Result<(Vec<u8>, usize), RecordStoreError> {
+        match self {
+            Self::InMemory(store) => store.serialize_slice(start),
+            Self::Rocksdb(store) => store.serialize_slice(start),
+        }
+    }
+
+    pub fn serialize_record(&self, record: &ProcessorRecord) -> Result<Vec<u8>, bincode::Error> {
+        let ProcessorRecord { values, lifetime } = record;
+        let values = values
+            .iter()
+            .map(|value| match (value, self) {
+                (RecordRef::InMemory(record_ref), ProcessorRecordStore::InMemory(record_store)) => {
+                    record_store.serialize_ref(record_ref)
+                }
+                (RecordRef::Rocksdb(record_ref), _) => *record_ref,
+                _ => panic!("In memory record ref cannot be serialized by rocksdb record store"),
+            })
+            .collect();
+        let record = ProcessorRecordForSerialization {
+            values,
+            lifetime: lifetime.clone(),
+        };
+        bincode::serialize(&record)
+    }
+}
+
+impl StoreRecord for ProcessorRecordStore {
+    fn create_ref(&self, values: &[Field]) -> Result<RecordRef, RecordStoreError> {
+        match self {
+            Self::InMemory(store) => Ok(RecordRef::InMemory(store.create_ref(values)?)),
+            Self::Rocksdb(store) => Ok(RecordRef::Rocksdb(store.create_ref(values)?)),
+        }
+    }
+
+    fn load_ref(&self, record_ref: &RecordRef) -> Result<Vec<Field>, RecordStoreError> {
+        match (record_ref, self) {
+            (RecordRef::InMemory(record_ref), _) => Ok(load_in_memory_record_ref(record_ref)),
+            (RecordRef::Rocksdb(record_ref), ProcessorRecordStore::Rocksdb(record_store)) => {
+                Ok(record_store.load_ref(record_ref)?)
+            }
+            _ => panic!("Rocksdb record ref cannot be loaded by in memory record store"),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum ProcessorRecordStoreDeserializer {
+    InMemory(in_memory::ProcessorRecordStoreDeserializer),
+    Rocksdb(rocksdb::ProcessorRecordStore),
+}
+
+impl ProcessorRecordStoreDeserializer {
+    pub fn new(record_store: RecordStore) -> Result<Self, RecordStoreError> {
+        match record_store {
+            RecordStore::InMemory => Ok(Self::InMemory(
+                in_memory::ProcessorRecordStoreDeserializer::new()?,
+            )),
+            RecordStore::Rocksdb => Ok(Self::Rocksdb(rocksdb::ProcessorRecordStore::new()?)),
+        }
+    }
+
+    pub fn deserialize_and_extend(&self, data: &[u8]) -> Result<(), RecordStoreError> {
+        match self {
+            Self::InMemory(store) => store.deserialize_and_extend(data),
+            Self::Rocksdb(store) => store.deserialize_and_extend(data),
+        }
+    }
+
+    pub fn deserialize_record(&self, data: &[u8]) -> Result<ProcessorRecord, RecordStoreError> {
+        let ProcessorRecordForSerialization { values, lifetime } = bincode::deserialize(data)?;
+        let mut deserialized_values = Vec::with_capacity(values.len());
+        for value in values {
+            match self {
+                Self::InMemory(record_store) => {
+                    let record_ref = record_store.deserialize_ref(value)?;
+                    deserialized_values.push(RecordRef::InMemory(record_ref));
+                }
+                Self::Rocksdb(_) => deserialized_values.push(RecordRef::Rocksdb(value)),
+            }
+        }
+        Ok(ProcessorRecord {
+            values: deserialized_values.into(),
+            lifetime,
+        })
+    }
+
+    pub fn into_record_store(self) -> ProcessorRecordStore {
+        match self {
+            Self::InMemory(record_store) => {
+                ProcessorRecordStore::InMemory(record_store.into_record_store())
+            }
+            Self::Rocksdb(record_store) => ProcessorRecordStore::Rocksdb(record_store),
+        }
+    }
+}
+
+impl StoreRecord for ProcessorRecordStoreDeserializer {
+    fn create_ref(&self, values: &[Field]) -> Result<RecordRef, RecordStoreError> {
+        match self {
+            Self::InMemory(store) => Ok(RecordRef::InMemory(store.create_ref(values)?)),
+            Self::Rocksdb(store) => Ok(RecordRef::Rocksdb(store.create_ref(values)?)),
+        }
+    }
+
+    fn load_ref(&self, record_ref: &RecordRef) -> Result<Vec<Field>, RecordStoreError> {
+        match (record_ref, self) {
+            (RecordRef::InMemory(record_ref), _) => Ok(load_in_memory_record_ref(record_ref)),
+            (
+                RecordRef::Rocksdb(record_ref),
+                ProcessorRecordStoreDeserializer::Rocksdb(record_store),
+            ) => Ok(record_store.load_ref(record_ref)?),
+            _ => panic!("Rocksdb record ref cannot be loaded by in memory record store"),
+        }
+    }
+}
+
+fn load_in_memory_record_ref(record_ref: &in_memory::RecordRef) -> Vec<Field> {
+    record_ref.load().iter().map(FieldRef::cloned).collect()
+}
+
+mod in_memory;
+mod rocksdb;
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(crate = "dozer_types::serde")]
+struct ProcessorRecordForSerialization {
+    values: Vec<u64>,
+    lifetime: Option<Box<Lifetime>>,
+}
 
 #[cfg(test)]
 mod tests {
-    use dozer_types::types::Field;
+    use std::time::Duration;
 
-    use crate::RecordRef;
+    use dozer_types::types::Timestamp;
 
-    #[test]
-    fn test_store_load() {
-        let fields = vec![
-            Field::String("asdf".to_owned()),
-            Field::Int(23),
-            Field::Null,
-            Field::U128(234),
-        ];
+    use super::*;
 
-        let record = RecordRef::new(fields.clone());
-        let loaded_fields: Vec<_> = record
-            .load()
-            .into_iter()
-            .map(|field| field.cloned())
-            .collect();
-        assert_eq!(&fields, &loaded_fields);
+    fn test_record() -> Record {
+        let mut record = Record::new(vec![
+            Field::Int(1),
+            Field::Int(2),
+            Field::Int(3),
+            Field::Int(4),
+        ]);
+        record.lifetime = Some(Lifetime {
+            reference: Timestamp::parse_from_rfc3339("2020-01-01T00:13:00Z").unwrap(),
+            duration: Duration::from_secs(10),
+        });
+        record
+    }
+
+    fn test_record_roundtrip_impl(record_store_kind: RecordStore) {
+        let record = test_record();
+        let record_store = ProcessorRecordStore::new(record_store_kind).unwrap();
+        let processor_record = record_store.create_record(&record).unwrap();
+        assert_eq!(record_store.load_record(&processor_record).unwrap(), record);
     }
 
     #[test]
-    fn test_ser_de() {
-        let fields = vec![
-            Field::String("asdf".to_owned()),
-            Field::Int(23),
-            Field::Null,
-            Field::U128(234),
-        ];
+    fn test_record_roundtrip() {
+        test_record_roundtrip_impl(RecordStore::InMemory);
+        test_record_roundtrip_impl(RecordStore::Rocksdb);
+    }
 
-        let record = RecordRef::new(fields.clone());
+    fn test_record_serialization_roundtrip_impl(record_store_kind: RecordStore) {
+        let record_store = ProcessorRecordStore::new(record_store_kind).unwrap();
+        let record = record_store.create_record(&test_record()).unwrap();
+        let serialized_record = record_store.serialize_record(&record).unwrap();
+        let data = record_store.serialize_slice(0).unwrap().0;
 
-        let bytes = dozer_types::bincode::serialize(&record).unwrap();
-        let deserialized: RecordRef = dozer_types::bincode::deserialize(&bytes).unwrap();
-        let loaded_fields: Vec<_> = deserialized
-            .load()
-            .into_iter()
-            .map(|field| field.cloned())
-            .collect();
-        assert_eq!(&fields, &loaded_fields);
+        let record_store = ProcessorRecordStoreDeserializer::new(record_store_kind).unwrap();
+        record_store.deserialize_and_extend(&data).unwrap();
+        let deserialized_record = record_store.deserialize_record(&serialized_record).unwrap();
+        assert_eq!(deserialized_record, record);
+    }
+
+    #[test]
+    fn test_record_serialization_roundtrip() {
+        test_record_serialization_roundtrip_impl(RecordStore::InMemory);
+        // TODO: enable this test when serialization is implemented for rocksdb
+        // test_record_serialization_roundtrip_impl(RecordStore::Rocksdb);
     }
 }
