+++ dozer-storage/Cargo.toml
@@ -13,6 +13,7 @@ lmdb-rkv = "0.14.0"
 lmdb-rkv-sys = "0.11.2"
 page_size = "0.5.0"
 pin-project = "1.1.0"
+rocksdb = "0.21.0"
 tokio = "1.28.2"
 
 [dev-dependencies]
