+++ dozer-storage/src/errors.rs
@@ -31,6 +31,9 @@ pub enum StorageError {
     // Error forwarding
     #[error("Lmdb error: {0}")]
     Lmdb(#[from] lmdb::Error),
+
+    #[error("Rocksdb error: {0}")]
+    Rocksdb(#[from] rocksdb::Error),
 }
 
 #[derive(Debug, Error)]
